Moves common state machine methods to concern
