+++ app/models/contract.rb
@@ -16,6 +16,7 @@
 
 class Contract < ApplicationRecord
   include AASM
+  include HasStateMachine
 
   aasm do
     state :proposal, initial: true
@@ -84,17 +85,5 @@ class Contract < ApplicationRecord
 
     months = (end_date.year * 12 + end_date.month) - (start_date.year * 12 + start_date.month)
     (budget / months).to_f.round(2)
-  end
-
-  def next_event
-    aasm.events(permitted: true).first.name.to_s
-  end
-
-  def next_state
-    aasm.states(permitted: true).first.name.to_s
-  end
-
-  def self.with_status(status)
-    where(aasm_state: status)
   end
 end
